// src/cache/Enums/cache.enum.ts
export enum CACHE_TYPE {
  ACTIVE_USER = 'AN',
  TEAM_NODE = 'TN',
  UPLINE_USER = 'UP',
  NODEK_USER = 'NU',
  IMPORT_USER = 'IM',
  ADMIN_USER = 'AD',
  WEEK_GRAPH_DATA = 'WGD',
  MONTHLY_GRAPH_DATA = 'MGD',
  YEARLY_GRAPH_DATA = 'YGD',
  MONTHLY_TOKEN_GRAPH_DATA = 'MTGD',
  YEARLY_TOKEN_GRAPH_DATA = 'YTGD',
  WEEK_TOKEN_GRAPH_DATA = 'WTGD',
  FIRSTLINE_USER_DATA = 'FLUD',
  ALL_USER_SUPERNODE = 'ALL_USER_SUPERNODE',
  USER_HIERARCHY = 'USER_HIERARCHY',

  //Wallet Deposit Settings

  WALLET_DEPOSIT_SETTINGS = 'WDS',
  STATIC_CACHE_KEY_FOR_DEPOSIT_SETTING = 'cache_homnifi_wallet_deposit_settings',

  //Wallet Withdraw settings

  WALLET_WITHDRAW_SETTINGS = 'WWS',
  STATIC_CACHE_KEY_FOR_WITHDRAW_SETTING = 'cache_homnifi_wallet_withdraw_settings',

  //Wallet Special Swap settings

  WALLET_SPECIAL_SWAP_SETTINGS = 'WSSC',
  STATIC_CACHE_KEY_FOR_WALLET_SPECIAL_SWAP_SETTING = 'cache_homnifi_wallet_special_swap_settings',

  //Wallet Special Swap settings

  WALLET_SWAP_SETTINGS = 'WSC',

  STATIC_CACHE_KEY_FOR_WALLET_SWAP_SETTING = 'cache_homnifi_wallet_swap_settings',
  TOTAL_REWARD_RESULT = 'TRR',
  REWARD_CLAIMABLE_RESULT = 'RCR',
  PLATFORM_SETTINGS = 'PLSTGS',
  ALL_PRODUCTS = 'ALLPRDCTS',
  FAQS = 'FQ',
  USER_CHILDREN_COUNT = 'UCC',
  REWARDS_ANALYTICS = 'RWAN',
  // SuperNode
  SUPERNODE_CURRENT_LOSS = 'SNCL',
  SUPERNODE_TEAM_TOTAL_PRODUCTION = 'SNTTP',
  SUPERNODE_TEAM_TOTAL_PRODUCTION_DAYS = 'SNTTPD',
  SUPERNODE_TEAM_TREE = 'SNTT',
  SUPERNODE_TEAM_TREE_SEARCH = 'SNTTS',

  ACTIVE_USER_ELIGIBLE = 'AUE',
  ACTIVE_BR_USER_ELIGIBLE = 'ACTIVE_BR_USER_ELIGIBLE', // Active Base Referral User Eligibliity
  BG_USER_HIGEST_MACHINE = 'BG_USER_HIGEST_MACHINE', // Active Base Referral User Eligibliity

  LEADERBOARD_DATA = 'LBD',
  LEADERS_DATA_BY_TYPE = 'LDBT',
}